src/test/java/ru/job4j/LogLineTest.java
package ru.job4j;

import org.junit.Test;

import static org.hamcrest.Matchers.is;
import static org.junit.Assert.assertThat;

public class LogLineTest {
    @Test
    public void whenParseSingleLine() {
        var line = new LogLine().parse(
                "WARNING [threads [zqfz [wbye [zfq]]] - 1] wqqa [e] zqw"
        );
        assertThat(line.level(),
                is("WARNING"));
        assertThat(line.threads(),
                is("[threads [zqfz [wbye [zfq]]] - 1]"));
        assertThat(line.text(),
                is("wqqa [e] zqw"));
    }

    @Test
    public void whenLevelError() {
        var line = new LogLine().parse(
                "ERROR [threads [wcxf [czed [ayya [fye]]]] - 0] eaba yzd"
        );
        assertThat(line.level(),
                is("ERROR"));
        assertThat(line.threads(),
                is("[threads [wcxf [czed [ayya [fye]]]] - 0]"));
        assertThat(line.text(),
                is("eaba yzd"));
    }

    @Test
    public void whenLevelInfo() {
        var line = new LogLine().parse(
                "INFO [threads [zqfz [wbye [zfq]]] - 1] qdcq"
        );
        assertThat(line.level(),
                is("INFO"));
        assertThat(line.threads(),
                is("[threads [zqfz [wbye [zfq]]] - 1]"));
        assertThat(line.text(),
                is("qdcq"));
    }
}